tool - split in multiple files
