+++ tools/lib/parse-package-json.js
@@ -1,0 +1,8 @@
+const fs = require('fs');
+
+module.exports = function getPackageConfig(configPath) {
+	const packageAsString = fs.readFileSync(configPath).toString();
+	const packageConfig = JSON.parse(packageAsString);
+
+	return packageConfig;
+}
